requirements.txt
confluent-kafka[avro,json,protobuf]
python-dotenv
pandas
